pyproject.toml
[tool.poetry]
name = "iiev-ultra"
version = "0.1.0"
description = "Invoice Ingestion and Validation Engine - Ultra Edition"
authors = ["IIEV Development Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
celery = "^5.3.4"
redis = "^5.0.1"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
pyodbc = "^5.0.1"
# psycopg2-binary = "^2.9.9"  # Nicht mehr benötigt - verwenden MSSQL
azure-storage-blob = "^12.19.0"
azure-servicebus = "^7.11.4"
azure-identity = "^1.15.0"
azure-keyvault-secrets = "^4.7.0"
lxml = "^4.9.3"
PyPDF2 = "^3.0.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "azure.*",
    "pyodbc.*",
    "celery.*"
]
ignore_missing_imports = true
